requirements.txt
aistudio-sdk==0.3.8
annotated-types==0.7.0
anyio==4.12.0
bce-python-sdk==0.9.57
certifi==2025.11.12
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
filelock==3.20.1
fsspec==2025.12.0
future==1.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.3
idna==3.11
imagesize==1.4.1
modelscope==1.33.0
networkx==3.6.1
numpy==2.2.6
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0.88
opt-einsum==3.3.0
packaging==25.0
paddleocr==3.3.2
paddlepaddle==3.2.2
paddlex==3.3.12
pandas==2.3.3
pillow==12.0.0
prettytable==3.17.0
protobuf==6.33.2
psutil==7.2.1
py-cpuinfo==9.0.0
pyclipper==1.4.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pypdfium2==5.2.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.11.3
requests==2.32.5
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
safetensors==0.7.0
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
tqdm==4.67.1
typer-slim==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.2
wcwidth==0.2.14